requirement.txt
pandas
scikit-learn
simpletransformers
tensorboardx
tensorflow
torch
numpy
kornia
openprompt